requirements.txt
streamlit
opencv-python
pytesseract
torch
ultralytics
mysql.connector
google.generativeai
Pillow 